Tic Tac Toe
A normal tic tac toe game where you play against the computer. The goal of this project is to be better at JavaScript but also trying to make a good A.I.

The goal of the A.I is to always check if it can win and then after that always try to stop the player from winning the the player have two in one row.